test(ui): split duplicate pane label coverage
